[bug #159006983] fix migration error
- delete the empty migration files

- create the migration file to include the new columns to the database
