--- alembic/versions/1574e2380537_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 1574e2380537
-Revises: 1662a541a209
-Create Date: 2018-07-10 12:05:20.949592
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '1574e2380537'
-down_revision = '1662a541a209'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
