--- alembic/versions/1662a541a209_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 1662a541a209
-Revises: 26c14e799a0f
-Create Date: 2018-07-09 13:30:26.239219
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '1662a541a209'
-down_revision = '26c14e799a0f'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
