+++ alembic/versions/5a0fcbced74d_add_the_country_and_timezone_attributes.py
@@ -1,0 +1,45 @@
+"""add the country and timezone attributes
+
+Revision ID: 5a0fcbced74d
+Revises: 26c14e799a0f
+Create Date: 2018-07-12 15:26:56.191582
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+
+# revision identifiers, used by Alembic.
+revision = '5a0fcbced74d'
+down_revision = '26c14e799a0f'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column('devices', 'location',
+               existing_type=sa.VARCHAR(),
+               nullable=False)
+    country_type = postgresql.ENUM('Uganda', 'Kenya', 'Nigeria', name='country_type')
+    country_type.create(op.get_bind())
+    timezone_type = postgresql.ENUM('EAST_AFRICA_TIME', 'WEST_AFRICA_TIME', name='timezone_type')
+    timezone_type.create(op.get_bind())
+    op.add_column('locations', sa.Column('country', sa.Enum('Uganda', 'Kenya', 'Nigeria', name='country_type'), nullable=True))
+    op.add_column('locations', sa.Column('time_zone', sa.Enum('EAST_AFRICA_TIME', 'WEST_AFRICA_TIME', name='timezone_type'), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('locations', 'time_zone')
+    op.drop_column('locations', 'country')
+    country_type = postgresql.ENUM('Uganda', 'Kenya', 'Nigeria', name='country_type')
+    country_type.drop(op.get_bind())
+    timezone_type = postgresql.ENUM('EAST_AFRICA_TIME', 'WEST_AFRICA_TIME', name='timezone_type')
+    timezone_type.drop(op.get_bind())
+    op.alter_column('devices', 'location',
+               existing_type=sa.VARCHAR(),
+               nullable=True)
+    # ### end Alembic commands ###
